backend_foodgram/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
django-filter==23.2
djoser==2.1.0
django-colorfield==0.10.1
reportlab==4.0.4
gunicorn==20.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
drf-extra-fields==3.7.0